Add working of roles to endpoints and add test for that

+++ api/users/handler.py
@@ -8,7 +8,7 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from api.auth.services.AuthService import AuthService
-from api.users.actions import activate_user_action
+from api.users.actions import activate_user_action, fetch_user_or_raise, check_user_permissions
 from api.users.actions import create_new_user_action
 from api.users.actions import delete_user_action
 from api.users.actions import get_user_by_id_action
@@ -21,6 +21,7 @@ from api.users.models import UpdateUserRequest
 from api.users.models import UserCreate
 from db.models import User
 from db.session import get_session
+from utils.roles import PortalRole
 
 logger = getLogger(__name__)
 
@@ -51,6 +52,16 @@ async def delete_user(
     current_user: User = Depends(AuthService.get_current_user_from_access_token),
     session: AsyncSession = Depends(get_session),
 ) -> DeleteUserResponse:
+    target_user = await fetch_user_or_raise(user_id, current_user.roles, session)
+    if target_user == current_user and PortalRole.ROLE_PORTAL_SUPERADMIN in current_user.roles:
+        raise HTTPException(status_code=406, detail="Superadmin cannot be deleted via API.")
+    
+    if not await check_user_permissions(
+        target_user=target_user,
+        current_user=current_user
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
+
     deleted_user_id = await delete_user_action(user_id, session)
     if deleted_user_id is None:
         raise HTTPException(
@@ -65,11 +76,14 @@ async def activate_user(
     current_user: User = Depends(AuthService.get_current_user_from_access_token),
     session: AsyncSession = Depends(get_session),
 ) -> ActivateUserResponse:
+    target_user = await fetch_user_or_raise(user_id, current_user.roles, session)
+    if not await check_user_permissions(
+        target_user=target_user,
+        current_user=current_user
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
+
     activated_user_id = await activate_user_action(user_id, session)
-    if activated_user_id is None:
-        raise HTTPException(
-            status_code=404, detail=f"User with id {user_id} not found."
-        )
     return ActivateUserResponse(activated_user_id=activated_user_id)
 
 
@@ -79,17 +93,19 @@ async def get_user_by_id(
     current_user: User = Depends(AuthService.get_current_user_from_access_token),
     session: AsyncSession = Depends(get_session),
 ) -> ShowUser:
-    user = await get_user_by_id_action(user_id, session)
-    if user is None:
-        raise HTTPException(
-            status_code=404, detail=f"User with id {user_id} not found."
-        )
+    target_user = await fetch_user_or_raise(user_id, current_user.roles, session)
+    if not await check_user_permissions(
+        target_user=target_user,
+        current_user=current_user
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
+
     return ShowUser(
-        user_id=user.user_id,
-        name=user.name,
-        surname=user.surname,
-        email=user.email,
-        is_active=user.is_active,
+        user_id=target_user.user_id,
+        name=target_user.name,
+        surname=target_user.surname,
+        email=target_user.email,
+        is_active=target_user.is_active,
     )
 
 
@@ -100,6 +116,13 @@ async def update_user_by_id(
     current_user: User = Depends(AuthService.get_current_user_from_access_token),
     session: AsyncSession = Depends(get_session),
 ) -> UpdatedUserResponse:
+    target_user = await fetch_user_or_raise(user_id, current_user.roles, session)
+    if not await check_user_permissions(
+        target_user=target_user,
+        current_user=current_user
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
+
     try:
         updated_user_id = await update_user_action(user_id, body, session)
     except IntegrityError as err:
